Add files via upload
Create a static method called convert that takes in two parameters, hours and minutes which are both integers. Inside the static method, write code to convert both hours and minutes to seconds. Multiply the number of hours by 3600 (the number of seconds in an hour) and the number of minutes by 60 (the number of seconds in a minute). Add these values together and return the total number of seconds.

Declare the following test cases in the Main() method:

Console.WriteLine(convert(2, 30)); // Expected output: 9000
Console.WriteLine(convert(1, 15)); // Expected output: 4500