.Split
Create a string variable called sentence and assign the value "The anemone, the wild violet, the hepatica, and the funny little curled-up." to it. Convert the string into an array. Use a loop to count the number of occurrences of the word "the" in the array.

In my solution, I used the String.Split static method to split the sentence into words by the space character.